+++ sdk/cognitivelanguage/azure-ai-language-questionanswering/azure/ai/language/questionanswering/models/_models.py
@@ -766,8 +766,7 @@ class PrebuiltQueryMatchingPolicy(MatchingPolicy, discriminator="Prebuilt"):
         """
 
     def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-        self.kind = MatchingPolicyKind.PREBUILT  # type: ignore
+        super().__init__(*args, kind=MatchingPolicyKind.PREBUILT, **kwargs)
 
 
 class QueryFilters(_Model):